# Improve performance of toMap when keys are mostly sorted