--- a/proto3-wire.cabal
+++ b/proto3-wire.cabal
@@ -65,3 +65,11 @@ test-suite tests
                        text >= 0.2 && <1.3,
                        transformers >=0.5.6.2 && <0.6,
                        vector >=0.12.0.2 && <0.13
+
+benchmark bench
+  type:                exitcode-stdio-1.0
+  main-is:             Main.hs
+  build-depends:       base >= 4 && < 5, bytestring, random, criterion, proto3-wire
+  hs-source-dirs:      bench
+  ghc-options:         -O2 -Wall -fobject-code -ddump-simpl -ddump-to-file
+